faster

+++ matlab/adaboost/ada_haar_define.m
@@ -52,9 +52,9 @@ function HAAR = ada_haar_define(varargin)
 IMSIZE = varargin{1};  IMAGE_W = IMSIZE(1); IMAGE_H = IMSIZE(2);  V = 0; D = 0;
 MAX_W = 15; MAX_H = 15; U_MIN_W = 1; U_MIN_H = 1; U_STEP_W = 1; U_STEP_H = 1;
 BIG_SAFE_NUMBER = 100000;                           % large enough for IMSIZE = [24 24]
-HAAR(BIG_SAFE_NUMBER).descriptor = [];
-HAAR(BIG_SAFE_NUMBER).hinds = [];
-HAAR(BIG_SAFE_NUMBER).hvals = [];
+HAAR(BIG_SAFE_NUMBER).descriptor = single([]);
+HAAR(BIG_SAFE_NUMBER).hinds = single([]);
+HAAR(BIG_SAFE_NUMBER).hvals = single([]);
 %HAAR(BIG_SAFE_NUMBER).theta = [];
 
 %WEAK.descriptor = zeros([BIG_SAFE_NUMBER 17]);      % preallocate descriptor space
@@ -103,11 +103,11 @@ if ismember('vert2', SHAPES)
             for y = 1:IMAGE_H - h + 1
                 for x = 1:IMAGE_W - w + 1
                    
-                    HAAR(c_num).descriptor(1) = 1;  % TYPE 1 = haar vertical
-                    HAAR(c_num).descriptor(2:3) = [x y];
-                    HAAR(c_num).descriptor(4:5) = [x+w/2-1 y+h-1];
-                    HAAR(c_num).descriptor(6:7) = [x+w/2 y];
-                    HAAR(c_num).descriptor(8:9) = [x+w-1 y+h-1];
+                    HAAR(c_num).descriptor(1) = single(1);  % TYPE 1 = haar vertical
+                    HAAR(c_num).descriptor(2:3) = single([x y]);
+                    HAAR(c_num).descriptor(4:5) = single([x+w/2-1 y+h-1]);
+                    HAAR(c_num).descriptor(6:7) = single([x+w/2 y]);
+                    HAAR(c_num).descriptor(8:9) = single([x+w-1 y+h-1]);
             
                     if (x > 1) && (y > 1)
                         HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y-1, x-1)];
@@ -123,9 +123,9 @@ if ismember('vert2', SHAPES)
 
                     end
                     
-                    HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y+h-1, x+w/2-1) sub2ind(IMSIZE, y+h-1, x+w-1)];
-                    HAAR(c_num).hvals = [HAAR(c_num).hvals -2  1];
-                    HAAR(c_num).polarity = 1;
+                    HAAR(c_num).hinds = single([HAAR(c_num).hinds sub2ind(IMSIZE, y+h-1, x+w/2-1) sub2ind(IMSIZE, y+h-1, x+w-1)]);
+                    HAAR(c_num).hvals = single([HAAR(c_num).hvals -2  1]);
+                    HAAR(c_num).polarity = single(1);
                     HAAR(c_num).theta = 0;
                     
                     if D   %plot the classifier
@@ -156,11 +156,11 @@ if ismember('horz2', SHAPES)
             for y = 1:IMAGE_H - h +1
                 for x = 1:IMAGE_W - w +1
 
-                    HAAR(c_num).descriptor(1) = 2;  % TYPE 2 = haar horizontal
-                    HAAR(c_num).descriptor(2:3) = [x y];
-                    HAAR(c_num).descriptor(4:5) = [x+w-1 y+h/2-1];
-                    HAAR(c_num).descriptor(6:7) = [x y+h/2];
-                    HAAR(c_num).descriptor(8:9) = [x+w-1 y+h-1];
+                    HAAR(c_num).descriptor(1) = single(2);  % TYPE 2 = haar horizontal
+                    HAAR(c_num).descriptor(2:3) = single([x y]);
+                    HAAR(c_num).descriptor(4:5) = single([x+w-1 y+h/2-1]);
+                    HAAR(c_num).descriptor(6:7) = single([x y+h/2]);
+                    HAAR(c_num).descriptor(8:9) = single([x+w-1 y+h-1]);
         
                     if (x > 1) && (y > 1)
                         HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y-1, x-1)];
@@ -175,9 +175,9 @@ if ismember('horz2', SHAPES)
                         HAAR(c_num).hvals = [HAAR(c_num).hvals 1];
                     end
                     
-                    HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y+h/2-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w-1)];
-                    HAAR(c_num).hvals = [HAAR(c_num).hvals -2 1];
-                    HAAR(c_num).polarity = 1;
+                    HAAR(c_num).hinds = single([HAAR(c_num).hinds sub2ind(IMSIZE, y+h/2-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w-1)]);
+                    HAAR(c_num).hvals = single([HAAR(c_num).hvals -2 1]);
+                    HAAR(c_num).polarity = single(1);
                     HAAR(c_num).theta = 0;
 
                     if D % plot the classifier
@@ -206,13 +206,13 @@ if ismember('vert3', SHAPES)
             for y = 1:IMAGE_H - h + 1
                 for x = 1:IMAGE_W - w + 1
                    
-                    HAAR(c_num).descriptor(1) = 3;  % TYPE 3 = 3-rectangle haar vertical
-                    HAAR(c_num).descriptor(2:3) = [x y];
-                    HAAR(c_num).descriptor(4:5) = [x+w/3-1 y+h-1];
-                    HAAR(c_num).descriptor(6:7) = [x+w/3 y];
-                    HAAR(c_num).descriptor(8:9) = [x+2*w/3-1 y+h-1];
-                    HAAR(c_num).descriptor(10:11) = [x+2*w/3 y];
-                    HAAR(c_num).descriptor(12:13) = [x+w-1 y+h-1];
+                    HAAR(c_num).descriptor(1) = single(3);  % TYPE 3 = 3-rectangle haar vertical
+                    HAAR(c_num).descriptor(2:3) = single([x y]);
+                    HAAR(c_num).descriptor(4:5) = single([x+w/3-1 y+h-1]);
+                    HAAR(c_num).descriptor(6:7) = single([x+w/3 y]);
+                    HAAR(c_num).descriptor(8:9) = single([x+2*w/3-1 y+h-1]);
+                    HAAR(c_num).descriptor(10:11) = single([x+2*w/3 y]);
+                    HAAR(c_num).descriptor(12:13) = single([x+w-1 y+h-1]);
 
                     
                     if (x > 1) && (y > 1)
@@ -228,9 +228,9 @@ if ismember('vert3', SHAPES)
                         HAAR(c_num).hvals = [HAAR(c_num).hvals 2 -2 1];
                     end
                     
-                    HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y+h-1, x+w/3-1) sub2ind(IMSIZE, y+h-1, x+2*w/3-1) sub2ind(IMSIZE, y+h-1, x+w-1)];
-                    HAAR(c_num).hvals = [HAAR(c_num).hvals -2 2 -1];
-                    HAAR(c_num).polarity = 1;
+                    HAAR(c_num).hinds = single([HAAR(c_num).hinds sub2ind(IMSIZE, y+h-1, x+w/3-1) sub2ind(IMSIZE, y+h-1, x+2*w/3-1) sub2ind(IMSIZE, y+h-1, x+w-1)]);
+                    HAAR(c_num).hvals = single([HAAR(c_num).hvals -2 2 -1]);
+                    HAAR(c_num).polarity = single(1);
                     HAAR(c_num).theta = 0;
 
                     if D % plot the classifier
@@ -259,13 +259,13 @@ if ismember('horz3', SHAPES)
             for y = 1:IMAGE_H - h + 1
                 for x = 1:IMAGE_W - w + 1
                    
-                    HAAR(c_num).descriptor(1) = 4;  % TYPE 4 = 3-rectangle haar horizontal
-                    HAAR(c_num).descriptor(2:3) = [x y];
-                    HAAR(c_num).descriptor(4:5) = [x+w-1 y+h/3-1];
-                    HAAR(c_num).descriptor(6:7) = [x y+h/3];
-                    HAAR(c_num).descriptor(8:9) = [x+w-1 y+2*h/3-1];
-                    HAAR(c_num).descriptor(10:11) = [x y+2*h/3];
-                    HAAR(c_num).descriptor(12:13) = [x+w-1 y+h-1];
+                    HAAR(c_num).descriptor(1) = single(4);  % TYPE 4 = 3-rectangle haar horizontal
+                    HAAR(c_num).descriptor(2:3) = single([x y]);
+                    HAAR(c_num).descriptor(4:5) = single([x+w-1 y+h/3-1]);
+                    HAAR(c_num).descriptor(6:7) = single([x y+h/3]);
+                    HAAR(c_num).descriptor(8:9) = single([x+w-1 y+2*h/3-1]);
+                    HAAR(c_num).descriptor(10:11) = single([x y+2*h/3]);
+                    HAAR(c_num).descriptor(12:13) = single([x+w-1 y+h-1]);
 
                     if (x > 1) && (y > 1)
                         HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y-1, x-1)];
@@ -280,9 +280,9 @@ if ismember('horz3', SHAPES)
                         HAAR(c_num).hvals = [HAAR(c_num).hvals 1];
                     end
 
-                    HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y+h/3-1, x+w-1) sub2ind(IMSIZE, y+2*h/3-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w-1)];
-                    HAAR(c_num).hvals = [HAAR(c_num).hvals -2 2 1];
-                    HAAR(c_num).polarity = 1;
+                    HAAR(c_num).hinds = single([HAAR(c_num).hinds sub2ind(IMSIZE, y+h/3-1, x+w-1) sub2ind(IMSIZE, y+2*h/3-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w-1)]);
+                    HAAR(c_num).hvals = single([HAAR(c_num).hvals -2 2 1]);
+                    HAAR(c_num).polarity = single(1);
                     HAAR(c_num).theta = 0;
 
                     if D % plot the classifier
@@ -312,15 +312,15 @@ if ismember('checker', SHAPES)
             for y = 1:IMAGE_H - h + 1
                 for x = 1:IMAGE_W - w + 1
                    
-                    HAAR(c_num).descriptor(1) = 5;  % TYPE 5 = haar 4-rectangle
-                    HAAR(c_num).descriptor(2:3) = [x y];
-                    HAAR(c_num).descriptor(4:5) = [x+w/2-1 y+h/2-1];
-                    HAAR(c_num).descriptor(6:7) = [x+w/2 y];
-                    HAAR(c_num).descriptor(8:9) = [x+w-1 y+h/2-1];
-                    HAAR(c_num).descriptor(10:11) = [x+w/2 y+h/2];
-                    HAAR(c_num).descriptor(12:13) = [x+w-1 y+h-1];
-                    HAAR(c_num).descriptor(14:15) = [x y+h/2];
-                    HAAR(c_num).descriptor(16:17) = [x+w/2-1 y+h-1];
+                    HAAR(c_num).descriptor(1) = single(5);  % TYPE 5 = haar 4-rectangle
+                    HAAR(c_num).descriptor(2:3) = single([x y]);
+                    HAAR(c_num).descriptor(4:5) = single([x+w/2-1 y+h/2-1]);
+                    HAAR(c_num).descriptor(6:7) = single([x+w/2 y]);
+                    HAAR(c_num).descriptor(8:9) = single([x+w-1 y+h/2-1]);
+                    HAAR(c_num).descriptor(10:11) = single([x+w/2 y+h/2]);
+                    HAAR(c_num).descriptor(12:13) = single([x+w-1 y+h-1]);
+                    HAAR(c_num).descriptor(14:15) = single([x y+h/2]);
+                    HAAR(c_num).descriptor(16:17) = single([x+w/2-1 y+h-1]);
 
                     if (x > 1) && (y > 1)
                         HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y-1, x-1)];
@@ -335,9 +335,9 @@ if ismember('checker', SHAPES)
                         HAAR(c_num).hvals = [HAAR(c_num).hvals 2 -1];
                     end
                     
-                    HAAR(c_num).hinds = [HAAR(c_num).hinds sub2ind(IMSIZE, y+h/2-1, x+w/2-1) sub2ind(IMSIZE, y+h/2-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w/2-1) sub2ind(IMSIZE, y+h-1, x+w-1)];
-                    HAAR(c_num).hvals = [HAAR(c_num).hvals -4 2 2 -1];
-                    HAAR(c_num).polarity = 1;
+                    HAAR(c_num).hinds = single([HAAR(c_num).hinds sub2ind(IMSIZE, y+h/2-1, x+w/2-1) sub2ind(IMSIZE, y+h/2-1, x+w-1) sub2ind(IMSIZE, y+h-1, x+w/2-1) sub2ind(IMSIZE, y+h-1, x+w-1)]);
+                    HAAR(c_num).hvals = single([HAAR(c_num).hvals -4 2 2 -1]);
+                    HAAR(c_num).polarity = single(1);
                     HAAR(c_num).theta = 0;
 
                     % plot the classifier
